create linkText component and structure changes
