+++ src/app/components/Header/Header.tsx
@@ -1,0 +1,192 @@
+'use client';
+
+import { useState } from 'react';
+import {
+    Dialog,
+    DialogPanel,
+    Disclosure,
+    DisclosureButton,
+    DisclosurePanel,
+    Popover,
+    PopoverButton,
+    PopoverGroup,
+    PopoverPanel,
+} from '@headlessui/react';
+import {
+    ArrowPathIcon,
+    Bars3Icon,
+    ChartPieIcon,
+    CursorArrowRaysIcon,
+    FingerPrintIcon,
+    SquaresPlusIcon,
+    XMarkIcon,
+} from '@heroicons/react/24/outline';
+import { ChevronDownIcon, PhoneIcon, PlayCircleIcon } from '@heroicons/react/20/solid';
+import { LinkText } from './components/link-text/LinkText';
+
+type Product = {
+    name: string;
+    description: string;
+    href: string;
+    icon: React.ComponentType<React.SVGProps<SVGSVGElement>>;
+};
+
+type CallToAction = {
+    name: string;
+    href: string;
+    icon: React.ComponentType<React.SVGProps<SVGSVGElement>>;
+};
+
+const products: Product[] = [
+    { name: 'Analytics', description: 'Get a better understanding of your traffic', href: '#', icon: ChartPieIcon },
+    { name: 'Engagement', description: 'Speak directly to your customers', href: '#', icon: CursorArrowRaysIcon },
+    { name: 'Security', description: 'Your customers’ data will be safe and secure', href: '#', icon: FingerPrintIcon },
+    { name: 'Integrations', description: 'Connect with third-party tools', href: '#', icon: SquaresPlusIcon },
+    { name: 'Automations', description: 'Build strategic funnels that will convert', href: '#', icon: ArrowPathIcon },
+];
+
+const callsToAction: CallToAction[] = [
+    { name: 'Watch demo', href: '#', icon: PlayCircleIcon },
+    { name: 'Contact sales', href: '#', icon: PhoneIcon },
+];
+
+export function Header() {
+    const [mobileMenuOpen, setMobileMenuOpen] = useState<boolean>(false);
+
+    return (
+        <header className="bg-white w-full">
+            <nav aria-label="Global" className="mx-auto flex max-w-7xl items-center justify-between p-6 lg:px-8">
+                <div className="flex lg:flex-1">
+                    <a href="#" className="-m-1.5 p-1.5">
+                        <span className="sr-only">Your Company</span>
+                        <img alt="" src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600" className="h-8 w-auto" />
+                    </a>
+                </div>
+                <div className="flex lg:hidden">
+                    <button
+                        type="button"
+                        onClick={() => setMobileMenuOpen(true)}
+                        className="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-gray-700"
+                    >
+                        <span className="sr-only">Open main menu</span>
+                        <Bars3Icon aria-hidden="true" className="h-6 w-6" />
+                    </button>
+                </div>
+                <PopoverGroup className="hidden lg:flex lg:gap-x-12">
+                    <Popover className="relative">
+                        <PopoverButton className="flex items-center gap-x-1 rounded-lg px-3 py-2 text-base font-semibold leading-6 text-gray-900 hover:bg-gray-50">
+                            Product
+                            <ChevronDownIcon aria-hidden="true" className="h-5 w-5 flex-none text-gray-400" />
+                        </PopoverButton>
+
+                        <PopoverPanel
+                            transition
+                            className="absolute -left-8 top-full z-10 mt-3 w-screen max-w-md overflow-hidden rounded-3xl bg-white shadow-lg ring-1 ring-gray-900/5 transition data-[closed]:translate-y-1 data-[closed]:opacity-0 data-[enter]:duration-200 data-[leave]:duration-150 data-[enter]:ease-out data-[leave]:ease-in"
+                        >
+                            <div className="p-4">
+                                {products.map((item) => (
+                                    <div
+                                        key={item.name}
+                                        className="group relative flex items-center gap-x-6 rounded-lg p-4 text-base leading-6 hover:bg-gray-50"
+                                    >
+                                        <div className="flex h-11 w-11 flex-none items-center justify-center rounded-lg bg-gray-50 group-hover:bg-white">
+                                            <item.icon aria-hidden="true" className="h-6 w-6 text-gray-600 group-hover:text-indigo-600" />
+                                        </div>
+                                        <div className="flex-auto">
+                                            <a href={item.href} className="block font-semibold text-gray-900">
+                                                {item.name}
+                                                <span className="absolute inset-0" />
+                                            </a>
+                                            <p className="mt-1 text-gray-600">{item.description}</p>
+                                        </div>
+                                    </div>
+                                ))}
+                            </div>
+                            <div className="grid grid-cols-2 divide-x divide-gray-900/5 bg-gray-50">
+                                {callsToAction.map((item) => (
+                                    <a
+                                        key={item.name}
+                                        href={item.href}
+                                        className="flex items-center justify-center gap-x-2.5 p-3 text-base font-semibold leading-6 text-gray-900 hover:bg-gray-100"
+                                    >
+                                        <item.icon aria-hidden="true" className="h-5 w-5 flex-none text-gray-400" />
+                                        {item.name}
+                                    </a>
+                                ))}
+                            </div>
+                        </PopoverPanel>
+                    </Popover>
+
+                    <LinkText>Features</LinkText>
+                    <LinkText>Marketplace</LinkText>
+                    <LinkText>Company</LinkText>
+                </PopoverGroup>
+                <div className="hidden lg:flex lg:flex-1 lg:justify-end">
+                    <a href="#" className="text-base font-semibold leading-6 text-gray-900 hover:bg-gray-50">
+                        Log in <span aria-hidden="true">&rarr;</span>
+                    </a>
+                </div>
+            </nav>
+            <Dialog open={mobileMenuOpen} onClose={setMobileMenuOpen} className="lg:hidden">
+                <div className="fixed inset-0 z-10" />
+                <DialogPanel className="fixed inset-y-0 right-0 z-10 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10">
+                    <div className="flex items-center justify-between">
+                        <a href="#" className="-m-1.5 p-1.5">
+                            <span className="sr-only">Your Company</span>
+                            <img
+                                alt=""
+                                src="https://tailwindui.com/img/logos/mark.svg?color=indigo&shade=600"
+                                className="h-8 w-auto"
+                            />
+                        </a>
+                        <button
+                            type="button"
+                            onClick={() => setMobileMenuOpen(false)}
+                            className="-m-2.5 rounded-md p-2.5 text-gray-700"
+                        >
+                            <span className="sr-only">Close menu</span>
+                            <XMarkIcon aria-hidden="true" className="h-6 w-6" />
+                        </button>
+                    </div>
+                    <div className="mt-6 flow-root">
+                        <div className="-my-6 divide-y divide-gray-500/10">
+                            <div className="space-y-2 py-6">
+                                <Disclosure as="div">
+                                    <DisclosureButton className="group flex items-center gap-x-1 rounded-lg px-3 py-2 text-base font-semibold leading-6 text-gray-900 hover:bg-gray-50 w-full justify-between">
+                                        Product
+                                        <ChevronDownIcon aria-hidden="true" className="h-5 w-5 flex-none text-gray-400 group-data-[open]:rotate-180" />
+                                    </DisclosureButton>
+                                    <DisclosurePanel className="mt-2 space-y-2">
+                                        {[...products, ...callsToAction].map((item) => (
+                                            <DisclosureButton
+                                                key={item.name}
+                                                as="a"
+                                                href={item.href}
+                                                className="block rounded-lg py-2 pl-6 pr-3 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50"
+                                            >
+                                                {item.name}
+                                            </DisclosureButton>
+                                        ))}
+                                    </DisclosurePanel>
+                                </Disclosure>
+
+                                <LinkText>Features</LinkText>
+                                <LinkText>Marketplace</LinkText>
+                                <LinkText>Company</LinkText>
+
+                            </div>
+                            <div className="py-6">
+                                <a
+                                    href="#"
+                                    className=" block rounded-lg px-3 py-2.5 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50"
+                                >
+                                    Log in
+                                </a>
+                            </div>
+                        </div>
+                    </div>
+                </DialogPanel>
+            </Dialog>
+        </header>
+    );
+}
