+++ src/app/page.tsx
@@ -1,10 +1,10 @@
 import Image from 'next/image';
-import { Header } from './about/components/Header/Header';
+import { Header } from './components/Header/Header';
 
 export default function Home() {
     return (
         <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-            <Header/>
+            <Header />
             <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
                 <h1>Welcome bliat</h1>
                 <Image
